refactor: ci for preprocessor (#462)
