+++ now/constants.py
@@ -6,7 +6,7 @@ from now.utils import BetterEnum
 DEMO_DATASET_DOCARRAY_VERSION = '0.13.17'
 
 DOCKER_BFF_PLAYGROUND_TAG = '0.0.116-use-cas-2'
-NOW_PREPROCESSOR_VERSION = '0.0.71-all-public-6'
+NOW_PREPROCESSOR_VERSION = '0.0.72-refactor-ci-preprocessor'
 
 NOW_AUTH_EXECUTOR_VERSION = '0.0.1'
